move two admin option classes to Admin; propagate options within the ADBsD.Admin's GetDatabaseAdmin method
